feat(selection_mode): small word
This is a better selection mode than the current word regex.

It allows word travesal for words that appear in camelCase, PascalCase,
snake_case and UPPER_CASE.

Very useful for changing the word in a variable name in most programming
languages.

In Vim, the default would treat camelCase, PascalCase, snake_case and
UPPER_CASE as one word, rendering the word movement useless, users are
forced to use character movement instead.