perf(syntax-highlight): highlight visible content only (#2)
Previously, syntax highlighting caused tremendous lag during typing for large files (over 2k lines). 

This is a known issue of Tree-sitter, that's why [nvim-treesitter](https://github.com/nvim-treesitter/nvim-treesitter#modules) has options to allow the user to disable it when the file is too large.

I tried to compensate for this by running the syntax highlighting on another thread, however, the improvement is negligible. 

One day, it just occurred to me that why don't I just highlight the content on the screen? Although it sounds like this can cause a lot of incorrect highlighting, surprisingly it's not, most of the time the highlighting is correct. I think this is due to the display of outbound parent lines, which allows the Tree-sitter parser to better recover from errors.