feat(global mode): file navigation (#11)
This is for users to navigate between opened files. It's better than Vim's `ctrl+o`/`ctrl+i` as Vim's file history is linear (like browser history), but Tim's file history is an undo-tree-like structure, which allows users to go to files that are no longer in the current linear history.

For example, if I navigate to A, then B, then back to A, then C, I will still be able to go to B by switching the history branch (by pressing `u` or `d` after entering the File Navigation mode).