wip: hackish but functional
This code does allow for document symbol navigation when used against
the Gleam LSP sever. Other LSP servers w/DocumentSymbol support were not
tested.

This has a few issues:
  * It does not support Nested DocumentSymbols
  * We "fake" a path using "/". DocumentSymbols does not return a path
from the server. Faking "/" seems like a bad idea. What if that path
goes somewhere else in code, such as to `rm -rf`.
  * We have to change the UI (go_to_location) to handle a deficiency in
the LSP code. When we use this "fake" path, we query the active
component and get it's path instead.