feat(keymap): add previous/next buffer navigation
Will cycle through currently open file buffers. If you get to the end of
the buffer list, it will start over at the beginning. Likewise, if you
are at the start of the buffer list and go previous, it will cycle back
to the last buffer.

When you close a buffer with `ctrl+c` , it will no longer be part of the
buffer navigation.

Currently, this functionality is bound to `-` and `=`. Better keys may
be `tab` and `shift+tab` but that has a conflict currently with Go back
and Go forward.

This is different than previous/next opened files as that will cycle
through files that you opened at some point in time. Closing a file does
not remove it from the history stack.