wip: move command's behind the --run paramater
This is semi-working.

* `ki readme.md` -- works
* `ki --run grammar fetch` -- works
* `ki --run log` -- works

ki log -- will not display the log path but it does not open up the file
log either. I am assuming that clap is parsing it into a command and
thus removing it from the args to the default edit command.

I am not yet sure how to get around this.