feat(cli): make edit the default command (#488)
* Nest subcommands under `@`

This solves the issue filename-subcommand ambiguity almost completely, albeit not perfect.
This solution assumes the user never needs to open a file named `@`, and adding new subcommands
in the future will not break existing behaviour.

For example, if `ki map` means "Open a file named `map`" today, adding a new subcommand `map` will
change the behaviour of `ki map` if we do not nest the subcommands under a placeholder like `@`.

The changes in this commit allowed running subcommands under `@`, for example `ki @ log` or
`ki @ grammar fetch`, however `ki <filename>` is broken.

Co-authored-by: wongjiahau <hou32hou@gmail.com>