feat: quick editor jumping (#509)
The numbers 1-5 can be used to "tag" an editor with that number. When pressing the number:

If no editors are tagged with the number pressed, the current editor will be tagged with that number.
If the current editor is already tagged with the number pressed, the tag will be cleared.
If another editor is tagged with the number pressed, you will jump to that editor.
This provides a very quick and efficient way of managing the most important files to you for your given editing session. The idea was built from other editor's plugins such as Harpoon, Reach and Lasso for Neovim.