refactor(app): reduce file opening dispatches to OpenFile and OpenFileFromPathBuf
These two dispatches take a path, owner and focus parameter. There were
a few calls to this Dispatch that were very easy to determine what the
parameters should be but there were a few others that I am not too sure
on right now. Needs more investigation. I did take guesses so that Ki
will run.

In `app.rs#to_dispatches` there are calls from
`DispatchPrompt::OpenFile` to `Dispatch::OpenFile` . What is using
`DispatchPrompt`?

...I am setting it to `User` and `true` for now.

In `components/dropdown.rs#From<CanonicalizedPath> for <DropdownItem>`
there are calls to `Dispatch::OpenFile`. How are we to know if this is a
system or user buffer and if the file should be focused or not?

... I am setting it to `User` and `true` for now.

In `generate_recipes.rs:62` there is a call to `Dispatch::OpenFile`,
what should it be?

... I am setting it to `User` and `true` for now.

All other calls to `Dispatch::OpenFile` are in tests and the parameters
can be derived from the context.