keymap: remove toggling of Line, Token & Word
I find that to cause a lot of mistakes, because I couldn't reliably keep track of the
current selection mode.

There were a few instances where I wished to jump to a specific line by executing
Line -> Jump, but I was already in Line, so the selection mode changed to Line Full
and the Jump didn't work as expected because Line Full selections can contain leading spaces.