VS Code Plug-in (#630)
This PR allows Ki motions to be used in VS Code.

Most of the features of Ki are usable in VS Code except (non-exhaustive):
1. Alt-key features: these are not mapped because they are mostly related to window management, which I think is best left handled by VS Code
2. Keymap Legend: this can be implemented in the future

The foundation created by this PR can also be used by other applications in the future such as Zed Editor, JetBrains IDE, etc.

Huge thanks to Evelant for initiating this PR.


Co-authored-by: WJH <hou32hou@gmail.com>