feat(movement): Previous/Next (#682)
This commit introduces two new movements: Previous and Next, which allows us to combine some selection modes or movements together:

1. Combine Token and Token Fine into Token.
2. Combine Word and Word Fine into Word.
3. Integrate empty lines navigation into Line (previously only available in Line Full).
4. Integrate sibling nodes movement that does not skip anonymous node into Syntax (previously only available in Syntax Fine).

These combinations eliminates the need of switching between different selection modes when specific movements are needed.

For example, previously, if we want to move to empty lines when we are already in Line selection mode, we have to switch to Line Full to achieve that.

Keymap changes:
- Find (Local) Backward/Forward have been rebound from `y`/`p` to `[`/`]`
- Previous/Next is bound to `y`/`p`