keymap(LineFull): similarize movement with Line Trimmed (#1003)
Also, in this commit, the special hidden movements DeleteForward and DeleteBackward are removed,
because they are no longer needed, as all selection modes now uses Right/Left to Delete forward/backward.

Previously it was needed, because Line Full uses Up/Down for Delete, so it requires these special movements
to override the default Delete behavior.