refactor(config): Use dedicated type for theme serde (#1106)
Instead of implementing custom Serialize/Deserialize implementations on
Theme, this change adds new types and makes use of serde attributes to
deserialize/serialize with distinct types.

The reason for this change is to pave the way for custom themes, and to
potentially allow embedding default themes, by having more correct serde
derives on Theme itself.