refactor(themes): embed compressed zed themes into the binary
Instead of downloading zed themes and caching them, we can directly
embed them into the binary. This change also refactors ThemeDescriptors
a bit, which is a bit more ergonomic and helps prepare better for custom
theme support.

Why not convert to a Ki theme, then embed? Because this requires some
sort of build script or pre-build step that converts, and doing that
right now is arduous at best. Besides, we need a zed theme converter in
the binary anyway, since we've more or less decided that custom theme
schema will simply be the Zed theme schema.

Why compress the Zed themes? Because they are 10x the size (around 700K)
otherwise, whereas presently they are ~60K.

Why miniz_oxide? It's a tiny library and an extremely popular crate: at
present the rust standard library even depends on it.

Last but not least, this change alters the zed theme parsing from
serde_json5 to serde_json_lenient. This is because zed actually uses the
latter.