refactor(ts_queries): Embed highlight queries into the binary
In a similar fashion to embedding zed themes, we should bundle highlight
queries. This allows us to do away with cache directories entirely, and
allows us to remove ureq from our dependency tree.

Unlike themes, highlight queries are something we want to update
regularly, so vendoring directly is cumbersome. A submodule is a
convenient choice - updating it is trivial.

Encoding the highlight queries is a bit troublesome, since they aren't
JSON like the Zed theme files were. Instead, we do some custom encoding.

Why not dump it as a JSON file, and then compress? Because this forces
us to put a JSON serializer into our build dependencies. Bloating build
dependencies is unwise because when we are cross-compiling, or building
in release mode, we need to build our serializer multiple times. In
release mode this is because build dependencies are un-optimized by
default.